Move transform methods to a new class, update copyrights and add limit_request to nginx.conf files
