+++ app/src/main/java/com/nginx/image/configs/HealthCheckConfiguration.java
@@ -15,7 +15,7 @@ import java.util.HashMap;
  *  ResizerConfiguration.java
  *  PhotoResizer
  *
- *  Copyright © 2017 NGINX Inc. All rights reserved.
+ *  Copyright © 2018 NGINX Inc. All rights reserved.
  *
  *  This class extends {@link Configuration}
  */
