+++ nginx/nginx-fabric.conf.j2
@@ -13,6 +13,9 @@
     proxy_set_header Connection "";
     proxy_set_header Accept-Encoding "";
 
+    limit_req zone=moderateReqs burst=150;
+    limit_req_status 503;
+
   }
 {%- endmacro %}
 
@@ -38,6 +41,9 @@ http {
     client_max_body_size 30M;
     client_body_buffer_size 30M;
     default_type application/octet-stream;
+  
+    # Moderated delivery: the rate should be based on the server capacity. This value works for a quad-core development system
+    limit_req_zone $server_addr zone=moderateReqs:1m rate=4r/s;
 
     {% if log_to_syslog %}
         log_format json '{"service_name": "{{ service_log_name }}",'
