+++ nginx/nginx-plus-fabric.conf.j2
@@ -13,6 +13,9 @@
     proxy_set_header Connection "";
     proxy_set_header Accept-Encoding "";
 
+    limit_req zone=moderateReqs burst=150;
+    limit_req_status 503;
+
     {% if healthcheck %}
       health_check interval=2s fails=1 uri={{healthcheck_uri or "/"}};
     {% endif %}
@@ -39,6 +42,9 @@ http {
   resolver {{ resolver.host }} valid={{ resolver.ttl }};  #use local DNS and override TTL to whatever value makes sense
   resolver_timeout 2s;
   client_max_body_size 30M;
+  
+  # Moderated delivery: the rate should be based on the server capacity. This value works for a quad-core development system
+  limit_req_zone $server_addr zone=moderateReqs:1m rate=4r/s;
 
     {% if log_to_syslog %}
         log_format json '{"service_name": "{{ service_log_name }}",'
